Flips OpenGL screenshots right side up
Screenshots taken with the OpenGL renderer previously were upside
down because OpenGl's origin is the lower left corner as opposed to
the upper left corner. This patch adds an option to save_png() that
flips the saved image by simply parsing rows in the reverse order,
and sets this value to true for OpenGL screenshots.