Fix the visual placement of the fake_exit trap
The fake_exit object was rendered on a different position then the one
shown in the editor while placing it. Only the visually position was
wrong, the actual "kill area" was at the correct position.

This commit fixes the visual placement, when playing a level, to match
the one shown in the editor. The "kill area" remains untouched.

Most of the currently included levels using the fake_exit trap remain
unaffected and/or are now fixed visually.

Fixes Pinugs/pingus#120